ALFREDAPI-377 Set default query consistency to TRANSACTIONAL_IF_POSSIBLE
The default query consistency was changed to TRANSACTIONAL_IF_POSSIBLE
because exact search terms don't work for the execution of queries with
eventual consistency. By setting the query consistency to
TRANSACTIONAL_IF_POSSIBLE we make sure that queries are only executed
against solr if it is needed. The default query consistency of Alfresco
is also TRANSACTIONAL_IF_POSSIBLE.